+++ R/merge_cell_seg_files.R
@@ -6,8 +6,10 @@
 #' @param base_path Path to a directory containing files to merge. All
 #' elegible files in this directory will be merged.
 #' @param update_progress Callback function which is called with progress.
+#' @param recursive If TRUE, will find files in subdirectories of `base_path`.
 #' @export
-merge_cell_seg_files = function(base_path, update_progress=NULL) {
+merge_cell_seg_files = function(base_path, update_progress=NULL,
+                                recursive=FALSE) {
 
   if (is.null(update_progress))
     update_progress = function(x) cat(x, '\n')
@@ -15,7 +17,8 @@ merge_cell_seg_files = function(base_path, update_progress=NULL) {
   for (suffix in merge_suffixes)
   {
     # Get paths to the files matching suffix
-    files = list.files(base_path, pattern=suffix, full.names=TRUE)
+    files = list.files(base_path, pattern=suffix,
+                       full.names=TRUE, recursive=recursive)
     if (length(files) == 0) next
 
     update_progress(paste('Merging', length(files), suffix, 'files.'))
@@ -33,10 +36,10 @@ merge_cell_seg_files = function(base_path, update_progress=NULL) {
 }
 
 # Helper for merge addin computes the number of progress messages we will emit.
-merge_progress_count = function(base_path) {
+merge_progress_count = function(base_path, recursive) {
   # Count the number of files of each type
   counts = merge_suffixes %>%
-    purrr::map_int(~length(list.files(base_path, pattern=.x))) %>%
+    purrr::map_int(~length(list.files(base_path, pattern=.x, recursive=recursive))) %>%
     purrr::keep(~.x>0)
 
   # Progress advances once for each file and once for each file type